requirements.txt
# Core Data Processing
pyspark==3.4.0
pandas>=1.5.0
numpy>=1.24.0

# Workflow Management
apache-airflow==2.7.0

# Database Connectivity
mysql-connector-python>=8.0.0

# Streaming
kafka-python>=2.0.0

# Web Requests
requests>=2.28.0

# Development & Analysis
jupyter>=1.0.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0

# Additional useful packages
plotly>=5.0.0
scikit-learn>=1.2.0